Always send the (correct) CA certificate to the client.
If we don't explicitly set SSL_ca, SSL_ca_file, or SSL_ca_path,
IO::Socket::SSL will try to do something clever by default: if it
knows of a CA certificate whose distinguished name matches the issuer
of the server certificate, it will send that certificate to the
client.

However, the islandhack *server* inherits SSL_CERT_DIR/SSL_CERT_FILE
from the parent process, which is necessary in order for --update to
work correctly and securely.

Thus, if the islandhack CA has the same name as some other CA that
openssl knows about, it would send the *wrong* certificate to the
client.  (Most clients don't care about this, but Java, for one,
does.)  This was a particular problem when running islandhack
recursively, but could conceivably be a problem in other situations as
well.

In any event, it may be useful to be able to retrieve the correct CA
certificate from islandhack directly, so do this.