Add a test suite.
To test the client-facing functionality of islandhack, we use various
HTTP/FTP clients (wget and curl, as well as the Java and Python
standard libraries) to attempt to download files from an example cache
directory.  It should be straightforward to add other command-line
clients to this test suite.

To test the server-facing (--update) functionality of islandhack, we
run islandhack inside another islandhack.  (This also serves as a test
case for LWP::UserAgent as a client.)  In keeping with islandhack's
philosophy, this operates without any network access.