Increase key size to 2048 bits.
For some reason, RSA keys smaller than 2048 bits do not work with
current IO::Socket::SSL + OpenSSL by default.  islandhack will simply
say "TLS handshake failed" and the client will report an error due to
the unexpected disconnection.

The logic is convoluted, and OpenSSL's error messages are not helpful.
But I think what is happening is that OpenSSL considers the key
insecure (which, y'know, it is) and refuses to use it; and it might be
possible to override that by setting the "security level", but
IO::Socket::SSL doesn't appear to provide a way to do that.

(My original choice to use a small key size was deliberate - I tried
to use poor security practices so that "smart" clients would notice
and complain, whereas "dumb" clients, like the ones islandhack is made
to interact with, wouldn't care.  But there's a limit to how much work
it's worth doing to make something deliberately insecure.  :P)