When creating a Java keystore, use JKS format.
More Java absurdity.  Trusted certificate authorities have to be
specified in the form of a "key store" file.  For some unfathomable
reason, a key store file must always have a password, and keytool will
refuse to work if the password is not at least 6 characters long.

Of course, the password does absolutely nothing... at least when the
key store is written in the traditional "JKS" format.  But apparently,
starting with JDK 9, keytool's default behavior is to generate a
"PKCS11" format file instead.  Naturally, this results in the
extremely helpful error message:

  javax.net.ssl.SSLException: Unexpected error:
  java.security.InvalidAlgorithmParameterException: the trustAnchors
  parameter must be non-empty

Fortunately keytool supports the "-storetype" option, to revert to the
marginally-less-absurd older behavior, and that option works as far
back as OpenJDK 7 and that's as far as I care to investigate.

Some more information and other possible workarounds can be found at:
https://git.mikael.io/mikaelhg/broken-docker-jdk9-cacerts